packages/policy/pyproject.toml
[project]
name = "lean-langchain-policy"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["lean-langchain-schemas", "lean-langchain-protocol", "pydantic>=2.8", "pyyaml>=6.0"]
